+++ Bonmin/src/Algorithms/Branching/BonChooseVariable.cpp
@@ -659,7 +659,7 @@ namespace Bonmin
       }
       int numberFixed=0;
       if (results_.size() > 0) {
-        returnCode = doStrongBranching(solver, info, results_.size(), 1);
+        returnCode = doStrongBranching(solver, info, (int)results_.size(), 1);
         if (bb_log_level_>=3) {
           OsiObject ** obj = solver->objects();
           const char* stat_msg[] = {"NOTDON", "FEAS", "INFEAS", "NOFINI"};
